cradle mount from the side
Take Two at a Gebildet motor mount. The first one was ok, but offset the motor mount point too far to one side. This one will offset the centre of balance, but if KSP has me anything, I'm sure it can be fixed with braces.

Oh and the top hole is a bit off, needs to move up about 2mm